refactor: game直下を削除
