+++ src/components/coop/NNAndIcon.tsx
