+++ src/components/coop/frame/step1/ThinkingTheme.tsx
