+++ src/components/coop/frame/step3/InputHint.tsx
@@ -15,7 +15,7 @@ import {
 
 import { CustomTitleText } from "@/components/common/CustomTitleText";
 import { VSpacer } from "@/components/common/Spacer";
-import { Wait } from "@/components/game/coop/frame/Wait";
+import { Wait } from "@/components/coop/frame/Wait";
 
 import { BASE_URL } from "@/data/data";
