+++ src/components/wolf/frame/step11/Score.tsx
@@ -5,6 +5,9 @@ import { useRecoilValue } from "recoil";
 
 import { Button, Center, Text, VStack } from "@chakra-ui/react";
 
+import { VSpacer } from "@/components/common/Spacer";
+import { PointList } from "@/components/wolf/PointList";
+
 import { BASE_URL, IS_AUTO_REQUEST } from "@/data/data";
 
 import { RecoilOwner, RecoilRoom } from "@/store/Recoil";
@@ -13,20 +16,16 @@ import { HandleError } from "@/hooks/useError";
 import { FetchStep } from "@/hooks/useFetchStep";
 import { AutoHttpRequest } from "@/hooks/useHttpRequest";
 
-import { CustomTitleText } from "../../../../common/CustomTitleText";
-import { VSpacer } from "../../../../common/Spacer";
-
 type Props = {
-  theme: string;
-  answer: string;
-  isCorrect: boolean;
+  playerList: { nickname: string; particIcon: number; point: number }[];
   setStep: Dispatch<SetStateAction<number>>;
 };
 
-export const Result = ({ theme, answer, isCorrect, setStep }: Props) => {
+export const Score = ({ playerList, setStep }: Props) => {
   const router = useRouter();
   const owner = useRecoilValue(RecoilOwner);
   const room = useRecoilValue(RecoilRoom);
+
   const Initialize = () => {
     axios
       .post(BASE_URL + "initialize", {
@@ -34,7 +33,7 @@ export const Result = ({ theme, answer, isCorrect, setStep }: Props) => {
       })
       .then((res) => {
         if (res.status === 200) {
-          router.push("/wait");
+          router.push("/wolf-wait");
         }
       })
       .catch((err) => {
@@ -49,8 +48,8 @@ export const Result = ({ theme, answer, isCorrect, setStep }: Props) => {
         params: { roomId: room.id },
       })
       .then((res) => {
-        if (res.data != 7) {
-          router.push("/wait");
+        if (res.data != 11) {
+          router.push("/wolf-wait");
         }
       })
       .catch((err) => {
@@ -75,32 +74,47 @@ export const Result = ({ theme, answer, isCorrect, setStep }: Props) => {
     <>
       <Center>
         <VStack>
-          <VSpacer size={4} />
-          <CustomTitleText title="お題" text={theme} />
-          <VSpacer size={4} />
-          <CustomTitleText title="解答" text={answer} />
-          <VSpacer size={12} />
-          {isCorrect ? (
-            <Text fontSize="3xl">正解！おめでとう！</Text>
-          ) : (
-            <Text fontSize="3xl">残念！公開処刑！</Text>
-          )}
-          <VSpacer size={12} />
+          <VStack>
+            <Text fontSize={24}>現在のポイント</Text>
+            <PointList memberPointList={playerList} />
+          </VStack>
+          <VSpacer size={16} />
+
           {owner.isOwner ? (
-            <Button colorScheme="red" minW={64} minH={12} onClick={Initialize}>
+            <Button
+              minW={60}
+              minH={12}
+              colorScheme="red"
+              color="white"
+              onClick={Initialize}
+            >
               次へ
             </Button>
           ) : IS_AUTO_REQUEST ? (
-            <Button colorScheme="red" minW={64} minH={12} isDisabled={true}>
+            <Button
+              minW={60}
+              minH={12}
+              colorScheme="red"
+              color="white"
+              isDisabled={true}
+            >
               待機中
             </Button>
           ) : (
-            <Button colorScheme="red" minW={64} minH={12} onClick={handleNext}>
+            <Button
+              minW={60}
+              minH={12}
+              colorScheme="red"
+              color="white"
+              onClick={handleNext}
+            >
               更新
             </Button>
           )}
         </VStack>
       </Center>
+
+      <VSpacer size={20} />
     </>
   );
 };
