+++ src/components/wolf/frame/step7/ThemeResult.tsx
