+++ src/components/wolf/frame/step8/ChoiceWolf.tsx
@@ -8,7 +8,7 @@ import { useRecoilValue } from "recoil";
 import { Button, Center, HStack, Image, Text, VStack } from "@chakra-ui/react";
 
 import { VSpacer } from "@/components/common/Spacer";
-import { Wait } from "@/components/game/coop/frame/Wait";
+import { Wait } from "@/components/coop/frame/Wait";
 
 import { BASE_URL } from "@/data/data";
 
@@ -18,7 +18,7 @@ import { Vote } from "@/types/type";
 
 import { HandleError } from "@/hooks/useError";
 
-import { ChoiceWolfList } from "../../parts/ChoiceWolfList";
+import { ChoiceWolfList } from "../../ChoiceWolfList";
 
 type Props = {
   setStep: Dispatch<SetStateAction<number>>;
