+++ src/components/wolf/frame/step9/BanishPerson.tsx
